fix(memory): address wiki git review feedback
